# src/main/webui/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ConnectionMemory:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemoryType"
        nextCursor:
          type: string
        hasNext:
          type: boolean
    ConnectionSearch:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchHitType"
        nextCursor:
          type: string
        hasNext:
          type: boolean
    ConnectionTeamMember:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TeamMemberType"
        nextCursor:
          type: string
        hasNext:
          type: boolean
    ConnectionTribute:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TributeType"
        nextCursor:
          type: string
        hasNext:
          type: boolean
    CreateInviteRequest:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/TeamRole"
        maxUses:
          type: integer
          format: int32
        expiresAt:
          $ref: "#/components/schemas/Instant"
    CreateMemoryRequest:
      type: object
      properties:
        title:
          type: string
        bodyText:
          type: string
        privateMemory:
          type: boolean
        taggedIds:
          type: array
          items:
            type: string
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
        mediaIds:
          type: array
          items:
            type: string
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        brandColor:
          type: string
    CreateTopicRequest:
      type: object
      properties:
        title:
          type: string
    CreateTributeRequest:
      type: object
      properties:
        recipientId:
          $ref: "#/components/schemas/UUID"
        text:
          type: string
        anonymous:
          type: boolean
        privateTribute:
          type: boolean
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    InviteByEmailRequest:
      type: object
      properties:
        email:
          type: string
        role:
          $ref: "#/components/schemas/TeamRole"
    InviteType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        teamId:
          $ref: "#/components/schemas/UUID"
        code:
          type: string
        role:
          $ref: "#/components/schemas/TeamRole"
        maxUses:
          type: integer
          format: int32
        useCount:
          type: integer
          format: int32
        expiresAt:
          $ref: "#/components/schemas/Instant"
        email:
          type: string
    JoinTeamRequest:
      type: object
      properties:
        code:
          type: string
        nickname:
          type: string
        bio:
          type: string
...
    MediaType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        url:
          type: string
        mimeType:
          type: string
    MemoryType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        teamId:
          $ref: "#/components/schemas/UUID"
        writer:
          $ref: "#/components/schemas/TeamMemberType"
        title:
          type: string
        bodyText:
          type: string
        privateMemory:
          type: boolean
        tagged:
          type: array
          items:
            $ref: "#/components/schemas/TeamMemberType"
        pictures:
          type: array
          items:
            $ref: "#/components/schemas/MediaType"
        createdAt:
          $ref: "#/components/schemas/Instant"
...
    SearchHitType:
      type: object
      properties:
        type:
          type: string
        id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
        snippet:
          type: string
    TeamMemberType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        teamId:
          $ref: "#/components/schemas/UUID"
        userId:
          $ref: "#/components/schemas/UUID"
        nickname:
          type: string
        bio:
          type: string
        role:
          $ref: "#/components/schemas/TeamRole"
        avatar:
          $ref: "#/components/schemas/MediaType"
    TeamRole:
      type: string
      enum:
      - ADMIN
      - MEMBER
    TeamType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        brandColor:
          type: string
        coverMedia:
          $ref: "#/components/schemas/MediaType"
        revealTributes:
          type: boolean
        revealAt:
          $ref: "#/components/schemas/Instant"
        tributesRevealed:
          type: boolean
        yearbookTitle:
          type: string
        yearbookSubtitle:
          type: string
        yearbookDedication:
          type: string
        yearbookTheme:
          $ref: "#/components/schemas/YearbookTheme"
        yearbookShowMembers:
          type: boolean
        yearbookShowTributes:
          type: boolean
        yearbookShowCharacteristics:
          type: boolean
        yearbookShowMemories:
          type: boolean
        yearbookShowAwards:
          type: boolean
...
    TopicType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        teamId:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
    TributeType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        teamId:
          $ref: "#/components/schemas/UUID"
        writer:
          $ref: "#/components/schemas/TeamMemberType"
        recipient:
          $ref: "#/components/schemas/TeamMemberType"
        text:
          type: string
        anonymous:
          type: boolean
        privateTribute:
          type: boolean
        published:
          type: boolean
        pictures:
          type: array
          items:
            $ref: "#/components/schemas/MediaType"
        createdAt:
          $ref: "#/components/schemas/Instant"
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
    UpdateTeamSettingsRequest:
      type: object
      properties:
        brandColor:
          type: string
        coverMediaId:
          $ref: "#/components/schemas/UUID"
        revealTributes:
          type: boolean
        revealAt:
          $ref: "#/components/schemas/Instant"
    UpdateYearbookSettingsRequest:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        dedication:
          type: string
        theme:
          $ref: "#/components/schemas/YearbookTheme"
        showMembers:
          type: boolean
        showTributes:
          type: boolean
        showCharacteristics:
          type: boolean
        showMemories:
          type: boolean
        showAwards:
          type: boolean
    UpsertTeamMemberProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        bio:
          type: string
        avatarId:
          $ref: "#/components/schemas/UUID"
...
    YearbookCharacteristicType:
      type: object
      properties:
        title:
          type: string
        count:
          type: integer
          format: int32
    YearbookCommentType:
      type: object
      properties:
        text:
          type: string
        writer:
          type: string
        imageUrls:
          type: array
          items:
            type: string
    YearbookMemberType:
      type: object
      properties:
        nickname:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/YearbookCharacteristicType"
        tributes:
          type: array
          items:
            $ref: "#/components/schemas/YearbookTributeType"
    YearbookMemoryType:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        writer:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/YearbookCommentType"
    YearbookStandingType:
      type: object
      properties:
        nickname:
          type: string
        score:
          type: integer
          format: int32
    YearbookTheme:
      type: string
      enum:
      - CLASSIC
      - MODERN
      - SCRAPBOOK
      - MINIMAL
    YearbookTopicType:
      type: object
      properties:
        title:
          type: string
        standings:
          type: array
          items:
            $ref: "#/components/schemas/YearbookStandingType"
    YearbookTributeType:
      type: object
      properties:
        text:
          type: string
        writer:
          type: string
    YearbookType:
      type: object
      properties:
        teamId:
          $ref: "#/components/schemas/UUID"
        orgName:
          type: string
        teamName:
          type: string
        title:
          type: string
        subtitle:
          type: string
        dedication:
          type: string
        theme:
          $ref: "#/components/schemas/YearbookTheme"
        brandColor:
          type: string
        logoUrl:
          type: string
        coverMediaUrl:
          type: string
        showMembers:
          type: boolean
        showTributes:
          type: boolean
        showCharacteristics:
          type: boolean
        showMemories:
          type: boolean
        showAwards:
          type: boolean
        members:
          type: array
          items:
            $ref: "#/components/schemas/YearbookMemberType"
        memories:
          type: array
          items:
            $ref: "#/components/schemas/YearbookMemoryType"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/YearbookTopicType"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
...
paths:
...
  /api/teams:
    get:
      summary: List teams for the current user
      tags:
      - Teams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    post:
      summary: Create a team
      tags:
      - Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/join:
    post:
      summary: Join a team with an invite code
      tags:
      - Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTeamRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}:
    patch:
      summary: Update team settings
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    get:
      summary: Get a team
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/invites:
    post:
      summary: Create an invite code
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/invites/email:
    post:
      summary: Invite a member by email
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteByEmailRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/members:
    get:
      summary: List team members
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: after
        in: query
        schema:
          type: string
      - name: first
        in: query
        schema:
          type: integer
          format: int32
      - name: query
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTeamMember"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/members/me:
    get:
      summary: Current user's membership in a team
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/memories:
    get:
      summary: List memories
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: after
        in: query
        schema:
          type: string
      - name: first
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionMemory"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    post:
      summary: Create a memory
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/profile:
    patch:
      summary: Upsert current user's team profile
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTeamMemberProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/search:
    get:
      summary: Search within a team
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: after
        in: query
        schema:
          type: string
      - name: first
        in: query
        schema:
          type: integer
          format: int32
      - name: q
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionSearch"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/topics:
    get:
      summary: List topics
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    post:
      summary: Create a topic
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/tributes:
    get:
      summary: List tributes
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: after
        in: query
        schema:
          type: string
      - name: first
        in: query
        schema:
          type: integer
          format: int32
      - name: recipientId
        in: query
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTribute"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    post:
      summary: Create a tribute
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTributeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TributeType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/yearbook:
    get:
      summary: Assembled yearbook for online viewing
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearbookType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
  /api/teams/{id}/yearbook-settings:
    patch:
      summary: Update yearbook layout settings
      tags:
      - Teams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateYearbookSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
...
info:
  title: Yaadbuzz REST API
  version: 1.0.0
  description: "Auth, media, and yearbook domain REST APIs for Yaadbuzz."